# frontend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Semenq API specification - AI-powered medicine accessibility platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medicines:
    get:
      operationId: listMedicines
      tags: [medicines]
      summary: List medicines
      parameters:
        - name: query
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of medicines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicineList"
...
  /pharmacies/{id}/inventory:
    get:
      operationId: getPharmacyInventory
      tags: [pharmacies]
      summary: Get pharmacy inventory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /pharmacies/{id}/inventory/{itemId}:
    patch:
      operationId: updateInventoryItem
      tags: [pharmacies]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryUpdate"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /analytics/demand-forecast:
    get:
      operationId: getDemandForecast
      tags: [analytics]
      summary: Get demand forecast for pharmacy inventory
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: Demand forecast data, when available
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DemandForecastItem"
  /analytics/top-medicines:
    get:
      operationId: getTopMedicines
      tags: [analytics]
      summary: Get recorded top medicines
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Recorded sales rankings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopMedicine"
...
components:
  schemas:
...
    Medicine:
      type: object
      required: [id, name, genericName, category, manufacturer, price, imageUrl]
      properties:
        id: { type: integer }
        name: { type: string }
        genericName: { type: string }
        category: { type: string }
        manufacturer: { type: string }
        composition: { type: string }
        dosage: { type: string }
        price: { type: number }
        mrp: { type: number }
        imageUrl: { type: string }
        requiresPrescription: { type: boolean }
        description: { type: ["string", "null"] }
        sideEffects: { type: ["string", "null"] }
        storageConditions: { type: ["string", "null"] }
    MedicineList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Medicine"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    InventoryItem:
      type: object
      required: [id, medicineId, medicineName, quantity, price, expiryDate, stockStatus]
      properties:
        id: { type: integer }
        medicineId: { type: integer }
        medicineName: { type: string }
        genericName: { type: string }
        quantity: { type: integer }
        price: { type: number }
        mrp: { type: number }
        expiryDate: { type: string }
        batchNumber: { type: string }
        stockStatus: { type: string, enum: [inStock, lowStock, outOfStock] }
        reorderLevel: { type: integer }
        lastRestocked: { type: string }
    InventoryUpdate:
      type: object
      properties:
        quantity: { type: integer }
        price: { type: number }
        expiryDate: { type: string }
        reorderLevel: { type: integer }
...
    DemandForecastItem:
      type: object
      required: [medicineId, medicineName, currentStock, predictedDemand, reorderSuggestion, confidence, trend]
      properties:
        medicineId: { type: integer }
        medicineName: { type: string }
        genericName: { type: string }
        currentStock: { type: integer }
        predictedDemand: { type: integer }
        reorderSuggestion: { type: integer }
        confidence: { type: number }
        trend: { type: string, enum: [rising, stable, falling] }
        healthStatus: { type: string, enum: [healthy, warning, critical] }
        aiInsight: { type: string }
        daysUntilStockout: { type: ["integer", "null"] }
    TopMedicine:
      type: object
      required: [medicineId, medicineName, count, revenue]
      properties:
        medicineId: { type: integer }
        medicineName: { type: string }
        category: { type: string }
        count: { type: integer }
        revenue: { type: number }
        trend: { type: string, enum: [up, down, stable] }
        percentChange: { type: number }
...